Make Terminal encoding nullable to automatically select one
If the encoding is not set, JLine automatically selects one that works
well for the current platform. Usually, this is the default system
encoding, but on Windows it now uses UTF-8 to allow Unicode support
when using Terminal.output() or Terminal.input().

The encoding is exposed using Terminal.encoding() and should be used
for reading and writing using the provided streams.